requirements.txt
pandas
numpy==1.26.4
scikit-learn
joblib
pyarrow
matplotlib
seaborn
mlflow
evidently==0.7.9
slack_sdk
yagmail
oauth2client  # if Gmail uses OAuth
streamlit
boto3
s3fs
python-dotenv
psycopg2-binary